Commit Mauro 26/12 22h20,
Ajout des cles etrangeres dans package Model,

Ajout des fonctions permettant de recuperer List sur TOUTES les DAO ( potentielles fautes de frappe dans les requetes => relire )

+++ src/Model/Perfusion.java
@@ -15,11 +15,11 @@ public class Perfusion {
     /**
      * Quantity of medication delivered
      */
-    private int quantity;
+    private float quantity;
     /**
      * Duration of a perfusion
      */
-    private int time;
+    private float time;
 
     /**
      * Default constructor, defines no arguments
@@ -33,7 +33,7 @@ public class Perfusion {
      * @param quantity initial quantity of a perfusion
      * @param time initial time of a perfusion
      */
-    public Perfusion(int id, String type, int quantity, int time) {
+    public Perfusion(int id, String type, float quantity, float time) {
         this.id = id;
         this.type = type;
         this.quantity = quantity;
@@ -68,23 +68,23 @@ public class Perfusion {
      * Get quantity of a perfusion
      * @return perfusion quantity
      */
-    public int getQuantity() { return quantity; }
+    public float getQuantity() { return quantity; }
 
     /**
      * Set quantity of a perfusion
      * @param quantity new perfusion quantity
      */
-    public void setQuantity(int quantity) { this.quantity = quantity; }
+    public void setQuantity(float quantity) { this.quantity = quantity; }
 
     /**
      * Get time of a perfusion
      * @return perfusion time
      */
-    public int getTime() { return time; }
+    public float getTime() { return time; }
 
     /**
      * Set time of a perfusion
      * @param time new perfusion time
      */
-    public void setTime(int time) { this.time = time; }
+    public void setTime(float time) { this.time = time; }
 }
